ref(*) move lib code into pkg
