+++ cmd/helm/get.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	"github.com/codegangsta/cli"
-	"github.com/deis/helm-dm/format"
+	"github.com/deis/helm-dm/pkg/format"
 )
 
 func init() {
