+++ cmd/helm/pack.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/codegangsta/cli"
-	"github.com/deis/helm-dm/format"
+	"github.com/deis/helm-dm/pkg/format"
 	"github.com/kubernetes/deployment-manager/chart"
 )
