+++ pkg/kubectl/delete.go
